refactor: Move PluralSelector under modules/genericeditor/components & CLDR_PLURALS to core/utils/constants
