+++ translate/src/core/utils/constants.ts
@@ -1,5 +1,3 @@
-export { default as PluralSelector } from './components/PluralSelector';
-
 // List of available CLDR plural categories.
 export const CLDR_PLURALS: Array<string> = [
   'zero',
