+++ translate/src/modules/fluenteditor/components/rich/RichTranslationForm.tsx
@@ -13,8 +13,8 @@ import { Locale } from '~/context/locale';
 import type { Translation } from '~/core/editor';
 import * as editor from '~/core/editor';
 import { useSelectedEntity } from '~/core/entities/hooks';
-import { CLDR_PLURALS } from '~/core/plural';
 import { fluent } from '~/core/utils';
+import { CLDR_PLURALS } from '~/core/utils/constants';
 import { useAppDispatch, useAppSelector } from '~/hooks';
 import { usePluralExamples } from '~/hooks/usePluralExamples';
 import { useReadonlyEditor } from '~/hooks/useReadonlyEditor';
