+++ translate/src/modules/genericeditor/components/PluralSelector.css
