+++ translate/src/modules/genericeditor/components/PluralSelector.test.js
